[tempfile] process-id filter on deletion to support forks

+++ git-tempfile/src/lib.rs
@@ -2,7 +2,8 @@
 //!
 //! This crate installs signal handlers the first time its facilities are used.
 //! These are powered by [`signal-hook`] to get notified when the application is told to shut down
-//! using signals to assure these are deleted.
+//! using signals to assure these are deleted. The deletion is filtered by process id to allow forks to have their own
+//! set of tempfiles that won't get deleted when the parent process exits.
 //!
 //! As typical handlers for `TERMination` are installed on first use and effectively overriding the defaults, we install
 //! default handlers to restore this behaviour.
@@ -31,7 +32,7 @@ use tempfile::NamedTempFile;
 
 static SIGNAL_HANDLER_MODE: AtomicUsize = AtomicUsize::new(SignalHandlerMode::default() as usize);
 static NEXT_MAP_INDEX: AtomicUsize = AtomicUsize::new(0);
-static REGISTER: Lazy<DashMap<usize, Option<NamedTempFile>>> = Lazy::new(|| {
+static REGISTER: Lazy<DashMap<usize, Option<ForksafeTempfile>>> = Lazy::new(|| {
     for sig in signal_hook::consts::TERM_SIGNALS {
         // SAFETY: handlers are considered unsafe because a lot can go wrong. See `cleanup_tempfiles()` for details on safety.
         #[allow(unsafe_code)]
@@ -46,12 +47,18 @@ mod handler {
 
     /// Note that Mutexes of any kind are not allowed, and so aren't allocation or deallocation of memory.
     pub fn cleanup_tempfiles() {
+        let current_pid = std::process::id();
         for mut tempfile in REGISTER.iter_mut() {
-            if let Some(tempfile) = tempfile.take() {
-                let (file, temppath) = tempfile.into_parts();
-                std::fs::remove_file(&temppath).ok();
-                std::mem::forget(temppath); // leak memory to prevent deallocation
-                file.sync_all().ok();
+            if tempfile
+                .as_ref()
+                .map_or(false, |tf| tf.owning_process_id == current_pid)
+            {
+                if let Some(tempfile) = tempfile.take() {
+                    let (file, temppath) = tempfile.inner.into_parts();
+                    std::fs::remove_file(&temppath).ok();
+                    std::mem::forget(temppath); // leak memory to prevent deallocation
+                    file.sync_all().ok();
+                }
             }
         }
     }
@@ -86,6 +93,20 @@ pub struct Registration {
     id: usize,
 }
 
+struct ForksafeTempfile {
+    inner: NamedTempFile,
+    owning_process_id: u32,
+}
+
+impl From<NamedTempFile> for ForksafeTempfile {
+    fn from(inner: NamedTempFile) -> Self {
+        ForksafeTempfile {
+            inner,
+            owning_process_id: std::process::id(),
+        }
+    }
+}
+
 mod registration {
     use crate::{Registration, NEXT_MAP_INDEX, REGISTER};
     use std::{io, path::Path};
@@ -111,6 +132,7 @@ mod registration {
                     builder
                         .rand_bytes(0)
                         .tempfile_in(path.parent().expect("parent directory is present"))?
+                        .into()
                 }),
             ));
             Ok(Registration { id })
@@ -118,7 +140,7 @@ mod registration {
 
         pub fn new(containing_directory: impl AsRef<Path>) -> io::Result<Registration> {
             let id = NEXT_MAP_INDEX.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-            expect_none(REGISTER.insert(id, Some(NamedTempFile::new_in(containing_directory)?)));
+            expect_none(REGISTER.insert(id, Some(NamedTempFile::new_in(containing_directory)?.into())));
             Ok(Registration { id })
         }
 
@@ -127,7 +149,7 @@ mod registration {
         pub fn take(self) -> Option<NamedTempFile> {
             let res = REGISTER.remove(&self.id);
             std::mem::forget(self);
-            res.and_then(|(_k, v)| v)
+            res.and_then(|(_k, v)| v.map(|v| v.inner))
         }
     }
